D1 mini: Switch a lamp by a relay and MQTT
This example uses a D1 mini as MQTT client to switch a lamp via a relay and measure the current. If the D1 mini receives a topic relay1/set/relay with the content (payload) on, it switches the relay on, with off it switches the relay off and with payload toggle it toggles the relay.

The two LEDS show the system state:

DUO-LED red blinking: No connection to WiFi (WLAN)
DUO-LED red: MQTT error
DUO-LED green: MQTT connection to broker OK. The green LED blinks every 2.5 seconds to show, that the D1 mini is working.
Don't forget to change WiFi data to your network values in line
MqttClientKH2 client("..ssid..", "..password..","mqtt server name");